move config into its own file to clean up main plugin file
